Don't use cached commit data
it wasn't finding the image linked, because of this
