+++ src/common/delivery/goals/fetchGoalsOnCommit.ts
@@ -15,7 +15,7 @@
  */
 
 import { HandlerContext, logger } from "@atomist/automation-client";
-import { SdmGoalsForCommit } from "../../../typings/types";
+import { CommitForSdmGoal, SdmGoalFields, SdmGoalRepo, SdmGoalsForCommit } from "../../../typings/types";
 
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import { NoCacheOptions } from "@atomist/automation-client/spi/graph/GraphClient";
@@ -23,6 +23,7 @@ import * as _ from "lodash";
 import { SdmGoal } from "../../../ingesters/sdmGoalIngester";
 import { Goal } from "./Goal";
 import { goalCorrespondsToSdmGoal } from "./storeGoals";
+import * as stringify from "json-stringify-safe";
 
 export async function findSdmGoalOnCommit(ctx: HandlerContext, id: RemoteRepoRef, providerId: string, goal: Goal): Promise<SdmGoal> {
     const sdmGoals = await fetchGoalsForCommit(ctx, id, providerId);
@@ -60,4 +61,18 @@ export async function fetchGoalsForCommit(ctx: HandlerContext, id: RemoteRepoRef
     }
 
     return _.flatten(result.SdmGoal);
+}
+
+
+export async function fetchCommitForSdmGoal(ctx: HandlerContext, goal: SdmGoalFields.Fragment & SdmGoalRepo.Fragment): Promise<CommitForSdmGoal.Commit> {
+    const variables = {sha: goal.sha, repo: goal.repo.name, owner: goal.repo.owner, branch: goal.branch};
+    const result = await ctx.graphClient.query<CommitForSdmGoal.Query, CommitForSdmGoal.Variables>(
+        {
+            options: NoCacheOptions,
+            name: "CommitForSdmGoal",
+            variables: {sha: goal.sha, repo: goal.repo.name, owner: goal.repo.owner, branch: goal.branch}});
+    if (!result || !result.Commit || result.Commit.length === 0) {
+        throw new Error("No commit found for goal " + stringify(variables));
+    }
+    return result.Commit[0];
 }
