+++ src/handlers/events/delivery/FulfillGoalOnRequested.ts
@@ -28,6 +28,7 @@ import {
     StatusForExecuteGoal,
 } from "../../../typings/types";
 import { executeGoal } from "./verify/executeGoal";
+import { fetchCommitForSdmGoal } from "../../../common/delivery/goals/fetchGoalsOnCommit";
 
 export class FulfillGoalOnRequested implements HandleEvent<OnAnyRequestedSdmGoal.Subscription>,
     EventHandlerMetadata {
@@ -96,12 +97,3 @@ function convertForNow(sdmGoal: SdmGoalFields.Fragment, commit: CommitForSdmGoal
     };
 }
 
-async function fetchCommitForSdmGoal(ctx: HandlerContext, goal: SdmGoalFields.Fragment & SdmGoalRepo.Fragment): Promise<CommitForSdmGoal.Commit> {
-    const variables = {sha: goal.sha, repo: goal.repo.name, owner: goal.repo.owner, branch: goal.branch};
-    const result = await ctx.graphClient.query<CommitForSdmGoal.Query, CommitForSdmGoal.Variables>(
-        {name: "CommitForSdmGoal", variables: {sha: goal.sha, repo: goal.repo.name, owner: goal.repo.owner, branch: goal.branch}});
-    if (!result || !result.Commit || result.Commit.length === 0) {
-        throw new Error("No commit found for goal " + stringify(variables));
-    }
-    return result.Commit[0];
-}
